Conditionally run go-md2man
For some reason not actually explained in upstream commit
eab73f3d514bc515c1eeb19926cb781410fd5de9, the man page was committed
in troff format to git directly, and is now illegible gook.

A consequence of this is we no longer need to run go-md2man (the
invocation of which should have been in the upstream Makefile directly
anyways).